Fix file and directory names containing dots
Since the package is supposed to be a sub-package in a namespace used by
other packages too, the package name contained a dot. Naturally, this
lead to file and directory names containing dots too, which Python
doesn't play particularly well with.
But after fixing this, everything seems to be working.