Adjust supported Python versions via `setup.cfg`
Instead of manually modifying generated files, one can pick the Python
versions for which build environments should be generated via the
`setup.cfg` file. Doing that and then running `tox -e bootstrap`
generates the files correctly, as can be seen by the fact that the
superfluous empty line which I left in when editing got removed when
generating the files.