Fix some of the badge URLs
The distinction between the user `gnn`, under which I register most of
these services, and the organisation `oemof`, under which the repository
is listed on github, leads to problems with the badges. I have to
manually figure out which of the names is used for the service's
respective page for information on the repository and which one is used
for the badge.