Use `find_namespace_packages` in `setup.py`
This is necessary since the `oemof` directory must not contain an
`__init__.py` file as that could prevent the one contained in the main
`oemof` package from being loaded. Without an `__init__.py` file,
`find_packages` doesn't find sub-modules and -packages. As namespace
packages are a feature introduced with Python 3, the `setuptools`
shipped with Python 2 interpreters doesn't contain the
`find_namespace_packages` function. That means these interpreters can't
be used with this package, not even to build the documentation. That's
the reason I'm removing them.