Don't use developmental status in README.rst badge
The version in `README.rst` is used for a badge which counts how many
commits there are already since the last stable release. Since the `dev`
status marks a version as a moving target, it doesn't make sense to
compare anything against it. In other words, the latest released version
is always the one with no suffix, which is why the status suffix isn't
used when serializing the version specifier in the README.