Fix datapackage deserialization test
Not importing `oemof.tabular.datapackage` means that the
`from_datapackage` method doesn't get added to `EnergySystems`, which is
why we have to at least import the package. In order to not have an
unused import or use what looks like a magically appearing class method,
I'm also changing the test to using the explicit
`deserialize_energy_system` function.