Make `datapackage.py` a deeper module in a sub-package
The `datapackage` namespace gets promoted to be a package, now
containing the module `reading` retaining most of the original contents
of `datapackage` from when it was a module. The new package only
re-exports the `deserialize_energy_system` function and uses it to
monkey patch the constructor class method into `EnergySystem`.