Add generated API documentation RST files
These where generated via:

```
sphinx-apidoc --implicit-namespaces --no-toc -o docs/reference \
              src/oemof \
              'src/oemof/tabular/examples*'
```

This has to be done every time a new source file is introduced. Maybe we
can automate it. We'll see.

The `--no-toc` flag is necessary, because the `index.rst` file present
in the directory already fulfills that purpose.
The `--implicit-namespaces` flag is necessary because `src/oemof`
doesn't (and can never) contain an `__init__.py` file, so without the
flag, Sphinx wouldn't recognize `oemof` to be the root package's name.
Finally, the last argument is the exclude pattern, preventing Sphinx
from trying to generate API documentation for the any Python files found
in the `examples` package.