Use PEP 508 URL requirements
According to the [release notes][0], support for this got added in
version 18.1, although the [reference][1] still claims Pip doesn't
support it. I should probably file an issue and a PR for that.
Using this, we can replace `dependency_links`, the downside being that
`oemof-tabular` can not be installed via PyPI as long as we depend on
unreleased versions of packages.

[0]: https://pip.pypa.io/en/stable/news/#id58
[1]: https://github.com/pypa/pip/blob/69037caa0c260484595a9698485d7267a3dc2d98/docs/html/reference/pip_install.rst#id35