Fix loading the default configuration
The function used to load configuration files was renamed and the use
site here didn't get the memo. ;)
The commit also adds a regression test checking that an attempt at
loading the default configuration is done, and that it's done with the
correct function, if the need to do so arises.
In order to get this test loaded though, the patterns used to identify
Python files containing tests had to be updated in `setup.cfg`.