Fix `facade.Link`'s docstring
One line was beyond 79 characters, a hard PEP 8 limit for everything.
This prompted me to also look at other lines which where longer than 72
characters, a PEP 8 limit for comments and other free flowing text. I
reformatted the lines to stay within the character limit.
While at it, I also changed the first line of the docstring to be a
the regular sentence that it is.