Remove previous foreign key config
Now the automatic inferrence of the foreign keys makes this code
obsolete. "bus.csv" is a special resource which must be there. All
fields from all resources which are string will be compared to the
column "name" of the bus resource and if there is a match, a foreign key
will be established. For profiles it it the header of all the resources
saved under "sequences" folder which will be searched through